+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - neo-tree: override examples, show hidden in alternate style
 - plantuml: key mapping to call out to plantuml jar or binary
 - example-config: add new plugins & override existing plugin config 
+- clojure: which-key mappings for nvim-treesitter-sexp structural editing
 
 ## Changed
 - readme: expand on description and use